spark/requirements.txt
pandas
pymongo
python-dotenv
pyspark
requests